+++ py_directus/models/__init__.py
@@ -1,5 +1,5 @@
 from .base import DirectusModel
 from .directus import *
 
-__all__ = (BaseDirectusActivity, BaseDirectusRevision, BaseDirectusRole, BaseDirectusRoles, BaseDirectusUser,
-           BaseDirectusFile, BaseDirectusFolder, BaseDirectusVersion,)
+__all__ = (BaseDirectusActivity, BaseDirectusComment, BaseDirectusRevision, BaseDirectusRole, BaseDirectusRoles,
+           BaseDirectusUser, BaseDirectusFile, BaseDirectusFolder, BaseDirectusVersion,)